trying to implement table for search results

+++ src/invoice/Invoice.java
@@ -13,9 +13,9 @@ public class Invoice {
 	private String _customerName;
 	private String _shippingAddress;
 	private String _invoiceAddress;
-	private float _ust;
-	private float _gross; //brutto
-	private float _net;
+	private double _ust;
+	private double _gross; //brutto
+	private double _net;
 	private ArrayList<InvoiceElement> _articles;
 	
 	public int get_invoiceNumber() {
@@ -89,27 +89,27 @@ public class Invoice {
 		this._invoiceAddress = _invoiceAddress;
 	}
 	
-	public float get_ust() {
+	public double get_ust() {
 		return _ust;
 	}
 	
-	public void set_ust(float _ust) {
+	public void set_ust(double _ust) {
 		this._ust = _ust;
 	}
 	
-	public float get_gross() {
+	public double get_gross() {
 		return _gross;
 	}
 	
-	public void set_gross(float _gross) {
+	public void set_gross(double _gross) {
 		this._gross = _gross;
 	}
 	
-	public float get_net() {
+	public double get_net() {
 		return _net;
 	}
 	
-	public void set_net(float _net) {
+	public void set_net(double _net) {
 		this._net = _net;
 	}
